+++ sdmx-dl-cli/src/main/java/sdmxdl/cli/ListCommand.java
@@ -28,7 +28,8 @@ import java.util.concurrent.Callable;
         subcommands = {
                 ListSourcesCommand.class,
                 ListFlowsCommand.class,
-                ListConceptsCommand.class,
+                ListDimensionsCommand.class,
+                ListAttributesCommand.class,
                 ListCodesCommand.class,
                 ListAvailabilityCommand.class,
                 ListFeaturesCommand.class,