+++ sdmx-dl-cli/src/test/java/sdmxdl/cli/ListCommandTest.java
@@ -14,7 +14,7 @@ public class ListCommandTest {
         CommandWatcher watcher = CommandWatcher.on(cmd);
 
         assertThat(cmd.execute()).isEqualTo(CommandLine.ExitCode.OK);
-        assertThat(watcher.getOut()).isNotEmpty().contains("sources", "flows", "concepts", "codes", "features", "drivers");
+        assertThat(watcher.getOut()).isNotEmpty().contains("sources", "flows", "dimensions", "attributes", "codes", "features", "drivers");
         assertThat(watcher.getErr()).isEmpty();
     }
 }