+++ CHANGELOG.md
@@ -23,6 +23,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - ![CLI] Use option instead of property to enable curl backend
 - ![CLI] Remove cache folder option
 - ![CLI] Replace position column with index column in `list/concepts` command
+- ![CLI] Split command `list/concepts` into `list/dimensions` and `list/attributes`
 
 ### Fixed